+++ .github/workflows/tests.yml
@@ -26,7 +26,7 @@ jobs:
 
       - name: Lint with flake8
         run: |
-          flake8 okf_collab_docs --count --show-source --max-complexity=10 --max-line-length=127 --statistics
+          flake8 . --count --show-source --max-complexity=10 --max-line-length=127 --statistics
 
       - name: Test with pytest
         run: |
